// src/app/features/monthly-stmt-pnl/models/monthly-stmt-pnl-response.model.ts
export interface MonthlyStmtPnlResponse {
  no: string;
  expdate: string;
  expym: string;
  currency: string;
  datano: string;
  ttypename: string;
  stock: string;
  stockname: string;
  qty: string;
  avgprice: string;
  amt: string;
  marketprice: string;
  marketvalue: string;
  profit: string;
  profitrate: string;
  divamt: string;
  profitdiv: string;
  profitratediv: string;
}